New User's Guide to Lutris
I have written up a new user's guide for Lutris, aimed at new users to gaming on Linux. It would be greatly appreciated if this was accepted into the project.

If there are any issues with spelling, syntax or technical misinformation, please do not hesitate to let me know and I will correct it accordingly.

Thank you so much for all the amazing work you do for Linux and gaming in general.